requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
openai==1.3.5
google-generativeai==0.3.1
requests==2.31.0
python-dotenv==1.0.0
# For local Whisper (optional)
# torch==2.1.0
# whisper==1.1.10
# For Google Cloud TTS (optional)
# google-cloud-texttospeech==2.14.1